+++ spec/unit/honeybadger/backend/server_spec.rb
@@ -5,6 +5,7 @@ require 'honeybadger/config'
 describe Honeybadger::Backend::Server do
   let(:config) { Honeybadger::Config.new(logger: NULL_LOGGER, api_key: 'abc123') }
   let(:logger) { config.logger }
+  let(:payload) { double('Notice', to_json: '{}') }
 
   subject { described_class.new(config) }
 
@@ -14,6 +15,26 @@ describe Honeybadger::Backend::Server do
     it "returns the response" do
       stub_http
       expect(notify_backend).to be_a Honeybadger::Backend::Response
+    end
+
+    context "when payload has an api key" do
+      before do
+        allow(payload).to receive(:api_key).and_return('badgers')
+      end
+
+      it "passes the payload api key in extra headers" do
+        http = stub_http
+        expect(http).to receive(:post).with(anything, anything, hash_including({ 'X-API-Key' => 'badgers'}))
+        notify_backend
+      end
+    end
+
+    context "when payload doesn't have an api key" do
+      it "doesn't pass extra headers" do
+        http = stub_http
+        expect(http).to receive(:post).with(anything, anything, hash_including({ 'X-API-Key' => 'abc123'}))
+        notify_backend
+      end
     end
 
     context "when encountering exceptions" do
@@ -57,7 +78,7 @@ describe Honeybadger::Backend::Server do
     end
 
     def notify_backend
-      subject.notify(:notices, double('Notice', to_json: '{}'))
+      subject.notify(:notices, payload)
     end
   end
 end